+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unreleased
 
+- Add error message when `configure_opentelemetry` fails
+  from auto-instrumentation
+
 ## 0.12b0
 
 - Support not configuring a token